Copied FAZ Release 1 to trunk ============================= - Allegiance is built against a modern compiler, offering slightly increased speeds and better interfaces with modern hardware. 	Sounds may sound clearer, and visuals may seem crisper depending on your hardware and drivers. - SOVLogin has been removed from the login sequence, and its core functions have been written into ASGS and Allegiance itself. (Pook, radar) - Chat Logging has been written into Allegiance itself via the LogChat String Registry Key. Set its value to "1" to enable chat logging. (radar, Dogbones) - When marking yourself as AFK, you will no longer be marked as "back" after a game has ended. (Imago) - When logging in with ASGS, you no longer have the ability to modify your callsign within Allegiance which used to cause "Authentication=2" errors. (radar) - The in-base Ship Loadout screen no longer requires 800x600 resolution. When docking/launching, your monitor's resolution will no longer switch. (FullMark? KGJV??) - Missing "Game Information" was added to the Game Info screen (Esc, I). (KGJV? FullMark?) - The Teleport Probe's 1-in-6 chance of "Random destruction" has been removed. All teleport probes will always work, and no longer have any chance of detonating. (Dogbones, radar) - Additional events are fired behind the scenes, allowing more detailed gamelogs to be captured for historical and game-balancing purposes. (Tigereye) - The "Help" screen has been rewritten to include up-to-date help information
git-svn-id: file:///media/data/aa/dev/Allegiance-SVN/trunk@43 8c327b6c-a73e-8745-8738-111f1e176ae1

+++ src/AGC/AGCGameParameters.cpp
@@ -916,7 +916,7 @@ STDMETHODIMP CAGCGameParameters::get_GoalTeamMoney(AGCMoney* pVal)
 STDMETHODIMP CAGCGameParameters::put_GoalTeamMoney(AGCMoney Val)
 {
   XLock lock(this);
-  m_mp.fGoalTeamMoney = Val;
+  m_mp.fGoalTeamMoney = static_cast<float>(Val);
   return S_OK;
 }
